/var/run/slapd also has to be accessible to slapd
Sharing unix sockets via data-only containers is a viable way of communicating between processes; to do this, however, `/var/run/slapd` has to be on a volume from an external container. There is no guarantee that it will have the correct permissions, hence the need to `chown` it in eintrypoint script.